Port to girepository-2.0
This is an initial port of lgi to use girepository-2.0 introduced in the
GLib 2.80 release.

If using 2.85, it will use the new gi_repository_dup_default() function to
ensure the repository is using the same repository as applications and
libraries such as libpeas.

There is still some work to be done, as libpeas tests do not yet pass
using this version of lgi.