Only look for zero-terminated arrays when necessary
LGI would previously assume that code was zero-terminated if the
resolved length was found to be -1 when marshalling an array. It had no
contingencies for when a length could not be found and GI confirmed the
array was not zero-terminated. This commit changes the logic to assume
that, in these cases, the array is of length one.

This commit does not appear to change any of LGI's behaviour—all it does
is it prevents accessing memory which hasn't been allocated.