Fix CI
Tests had previously been changed to work with versions of GI-enabled
libraries from Debian Trixie, but those libraries are not yet in Ubuntu
24.04 LTS, which CI uses. There are a few exceptions to this restoration
compared to upstream LGI: GTK is switched to version 4 because some of
LGI's GTK3 overrides appear to be failing. Additionally, at-spi2-core is
added to CI dependencies so that the accessibility bus is made available
to CI, which is required by some tests. This fixes CI.

This commit also re-disables tests on non-master branches, as they
should no longer be needed.