Modify Gio's async to work with Gtk4 conventions
Previously, async functions in GObject libraries were always named
<name>_async, with their finisher functions named as <name>_finish.
Newer Gtk4 classes buck this trend by instead having functions named
<name> and <name>_finish. LuaGObject previously expected that in these
cases, there would always be a function named <name>_sync, which is not
true for Gtk4's FileDialog class. This commit tweaks the Gio.Async
override logic such that it is assumed that if the user is trying to
call async_<name>, it is always assumed that <name> is the async
initiator in the absence of <name>_async.